edge/analysis-api/requirements.txt
# Core — used by the Firebase-driven worker (src.worker).
numpy>=1.24,<2.0
torch==2.4.0
torchaudio
soundfile
batdetect2==1.3.1
psycopg2-binary
firebase-admin>=6.0,<7.0
google-cloud-storage>=2.10

# Dead deps — still wired up so the legacy FastAPI HTTP path
# (src.main) remains buildable. Flip the Dockerfile CMD back to
# `uvicorn src.main:app` to use it.
fastapi
uvicorn[standard]
python-multipart
librosa
transformers
scikit-maad
pandas
